requirements.txt
Django==5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
segno==1.6.1
typer==0.12.3
Pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
